% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{apply_per_group}
\alias{apply_per_group}
\alias{sapply_per_group}
\title{Apply functions}
\usage{
apply_per_group(baf, FUN, ..., by_s = NULL, by_b = NULL,
  passBAf = FALSE)

sapply_per_group(baf, FUN, ..., by_s = NULL, by_b = NULL,
  b_simplify = T, s_simplify = b_simplify)
}
\arguments{
\item{baf}{an object of \code{\link{BAf-class}}}

\item{FUN}{the function}

\item{...}{the arguments that will be passed to FUN}

\item{by_s, by_b}{by which the samples (\var{by_s}) or binders (\var{by_b})
are grouped. The \var{FUN} is applied to each group and the results are
combined. This can be given by 1) a \code{\link{character}} string of a
column name in \code{@sinfo} or \code{@binder}, and 2) a
\code{\link{factor}} vector that can be used in the stratification. If it
is NULL as the default, there will be no stratification. So, the \var{FUN}
will be applied to entire data.}

\item{passBAf}{if BAf object instead of matrix in @.Data is passed when
applying the \var{FUN}.}

\item{b_simplify}{if the output should be simplified like \code{simplify} for
\code{sapply}}

\item{s_simplify}{if the output within stratified group by the \code{by_s}
should be simplified.}
}
\value{
an object of \code{\link{BAf-class}}
}
\description{
\code{apply_per_group} : Apply a function to each group which is a subset of 
samples and binders stratified by the variables in \code{@sinfo} and 
\code{@binder}. The variable names are given by the argument \code{by_s} and
\code{by_b}.
}
\examples{
data(sba)
summary(sba)

# mean per plate and assay
apply_per_group(sba, mean, by_s= "plate", by_b= "assay")

# calculate median after stratifying samples by "plate" and "sex"
apply_per_group(sba, median, na.rm= TRUE, by_s= c("plate", "sex"))

# verify the first median computed by above line
median(sba@.Data[with(sba@sinfo, plate == "1" & !is.na(sex) & sex == "female"), ])

# compute coefficients of variation (CV) of \\code{MIX_1} wells
apply_per_group(sba[sba@sinfo$cohort == "MIX_1", 1:5], FUN= function(j) {
    apply(j, 2, function(k) sd(k, na.rm= TRUE)/mean(k, na.rm= TRUE))
})

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
