% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC.SBA.R
\name{ask_save.plot_QC.SBA}
\alias{ask_save.plot_QC.SBA}
\title{plot for basic QC of SBA data}
\usage{
ask_save.plot_QC.SBA(baf, prefix = deparse(substitute(baf)),
  path = "../output/QC", bead_count = NULL,
  sample_order = order(batch(baf, "sample")),
  color_tbl_b = default_color_SBA_bead(T),
  color_tbl_s = default_color_SBA_ctrl_samples(sid(baf)),
  main = paste(unique(batch(baf, "binder")), collapse = ", "),
  ylab_signal = "Raw MFI", guide_line = TRUE)
}
\arguments{
\item{baf}{an object of the \code{\link{BAf-class}} that contains SBA data}

\item{prefix}{a prefix of generating files of basic QC plots. Often, it is
sample layout ID.}

\item{path}{path where the plots will be saved}

\item{bead_count}{1) a \code{data.frame}, 2) a file name or 3) a
\code{matrix}. check \code{\link{plot_QC_bead_count_boxplot}}}

\item{sample_order}{the order of sample in plots. The default is the order by
sample batch in order to easily check batch differences.}

\item{color_tbl_b}{the colors of the binders employed for positive or 
negative control on plot. The information should be provided as a 
\code{\link{data.frame}}, in which columns are \code{name}, \code{reg}
(regular expression to find the binder), \code{col} (color), and
\code{type} (\code{positive} , \code{negative}, or \code{measure}). If this
is given as NULL, it results in no coloring.}

\item{color_tbl_s}{colors of samples on the plot. The information should be 
provided as a \code{\link{data.frame}} that has \code{name}, \code{reg} 
(regular expression to find the sample), \code{col} (color), \code{type} 
(\code{negative} or \code{replicated}).\cr If this is NULL, then no
coloring is applied.}

\item{main}{same as \code{\link{boxplot}}}

\item{ylab_signal}{the label of y-axis of the plots that show signal
distribution in boxplots.}

\item{guide_line}{if some guide lines of controls are displayed.}
}
\value{
\code{ask_save.plot_QC.SBA}
}
\description{
Generate plots to check basic QC of SBA data. This invokes other
\code{plot_QC} functions, such as 
\code{\link{plot_QC_sample_signal_boxplot}},
\code{\link{plot_QC_binder_signal_boxplot}}, \code{\link{plot_QC_repl_var}},
and \code{\link{plot_QC_bead_count_boxplot}}
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
