% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.BAf-class.R
\docType{class}
\name{BAf-class}
\alias{BAf-class}
\alias{BAf}
\title{An S4 class to handle the data from Binder Arrays}
\description{
This class is designed to store the data generated by Binder Arrays
(including suspension bead arrays) technology and to facilitate the analysis
of the data. It is distributed and stored in 5 main slots, \code{.Data},
\code{sinfo}, \code{binder}, \code{assy_s}, and \code{assy_b}. In order to
assure the orders of samples and targets are kept same in all the variables,
\code{'id'}, \code{colnames}, and \code{rownames} are checked.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\code{\link{matrix}} - row : samples, col : antibody\cr It 
contains all measured values. Its column names are reserved for target IDs.
The row names are the keys to link to \code{@sinfo} assuring the order of 
sample is same in both.}

\item{\code{sinfo}}{\code{\link{tbl_df}} - row : samples, col : variables, that
contains sample information\cr Each row is for each sample. Each column has
a type of information (e.g age) about the sample. Some description about
the columns can be found in the \code{@codebook}. Two columns in this table
are special.\cr 1) A column \code{'id'} is reserved for the unique
identifiers of the samples stored in a BAf-class object. The \code{'id'} is
used to confirm the order of the samples are same as other slots. 2) One
more column is reserved to indicate the batch of the samples. The name of
this column is given in \code{@batch_c}.}

\item{\code{binder}}{\code{\link{data.frame}} - row : binder (or antibodies / 
antigen), col : annotations for the binders\cr All information about beads 
or bound antibodies are stored here. Rows are in the same order of columns 
in \code{@.Data}.\cr 1) Like \code{@sinfo}, A column \code{'id'} is 
reserved for the unique identifiers of the binders, which have to be 
identical to column names of \code{@.Data}.\cr And one column should be 
assigned for binder batches. The name of the column is given in 
\code{@batch_c}.}

\item{\code{batch_c}}{\code{\link{list}} - two elements, \code{sinfo} and 
\code{binder}. Each has one text to indicate the column that contains batch
IDs for sample or binder in \code{@sinfo} or \code{@binder}.}

\item{\code{assy_s}}{\code{\link{list}} - each element is a \code{data.frame} of a 
particular batch-wise property (e.g. assay_failure). The columns of the 
\code{data.frame} are the binder batches, while rows are samples. In every 
data frame, the column names should be matched to the binder batch IDs in 
\code{@binder}.}

\item{\code{assy_b}}{\code{\link{list}} - Like \code{@assy_s}, but for binders. Each 
row is for each binder in \code{@binder}. The columns are for the sample 
batches.}

\item{\code{codebook}}{\code{\link{list}} - two elements, \code{sinfo} and
\code{binder}. Each contains a data frame of the codebook for \code{@sinfo}
or \code{@binder}.}

\item{\code{note}}{any}
}}

\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
