% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.BAf-package.R
\docType{package}
\name{BAf-package}
\alias{BAf-package}
\title{BAf : A package for handling the data from Binder Arrays in facility}
\description{
This package includes class 'BAf' and functions to assist in the
  analysis of data generated by Binder Arrays (including suspension bead
  arrays) technology in affinity proteomics facility.
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
\keyword{package}
