% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.cbind_rbind.R
\name{cbind_rbind}
\alias{cbind_rbind}
\alias{cbind.BAf}
\alias{rbind.BAf}
\title{Combine BAf objects by Rows or Columns}
\usage{
\method{cbind}{BAf}(..., deparse.level = 1)

\method{rbind}{BAf}(..., deparse.level = 1)
}
\arguments{
\item{...}{one or more objects of the \code{\link{BAf-class}}}

\item{deparse.level}{not used. a variable included in the generic function}
}
\value{
an object of the \code{\link{BAf-class}} after binding
}
\description{
Both functions works in almost same way as generic forms of them.\cr 
\itemize{ 
\item \code{cbind.BAf} binds BAf objects in column-wise. It checks if it is 
plausible to bind two or more BAf objects first, such as identical sample
information stored in \code{@sinfo} and sample batch-wide data in in
\code{@assy_s}.\cr
\item \code{rbind.BAf} binds in row-wise. To confirm it is possible to
combine multiple data, this checks the consistency in variable names in 
\code{@sinfo} and in \code{@assy_s}, the contents in \code{@assy_b} and in
\code{@binder}\cr
}
}
\examples{
data(sba)
summary(sba@binder)

sba2 <- cbind(sba[, 1:3], sba[, 10:20])
summary(sba2)

sba3 <- rbind(sba[1:5, ], sba[10:20, ])
summary(sba3)

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
