% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.combine_repeats.R
\name{combine_repeats}
\alias{combine_repeats}
\title{Function to combine repeated measures}
\usage{
combine_repeats(
  baf,
  rep_binder,
  rm_grp_rep_binder = T,
  center = T,
  scale = F,
  FUN = function(x) mean(x, na.rm = T),
  ...
)
}
\arguments{
\item{baf}{an object of the \code{\link{BAf-class}}}

\item{rep_binder}{a \code{factor} (or a \code{vector}) that indicates which
binders were repeated, so their values should be combined.}

\item{rm_grp_rep_binder}{if the \code{'grp'} and \code{'rep_binder'} columns
in \code{@binder} that were added by 
\code{\link{transform_to_landscape_for_repeats}}, they should be removed.}

\item{center, scale}{if the difference between assays should be adjusted using
\code{\link{scale}}}

\item{FUN}{the function to compute a reference value such as mean and median}

\item{...}{not used}
}
\value{
A \code{\link{BAf-class}} in which the two observed values were replaced with
the representative value computed by the \code{FUN}.
}
\description{
Merge repeated measures by taking one representative value (e.g. average) of 
them. It assumes the measurements were conducted in separate binder batches
(or assays), so that the data are stored in different columns of \code{baf}.
If repeated measurements were conducted in a same binder batch, transform the
data using \code{\link{transform_to_landscape_for_repeats}}.
}
\seealso{
\code{\link{transform_to_landscape_for_repeats}}
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
