% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.data_access.R
\name{data_access}
\alias{data_access}
\alias{sI}
\alias{sI,BAf-method}
\alias{sI<-}
\alias{sI<-,BAf-method}
\alias{sX}
\alias{sX,BAf-method}
\alias{sX<-}
\alias{sX<-,BAf-method}
\alias{sB}
\alias{sB,BAf-method}
\alias{sB<-}
\alias{sB<-,BAf-method}
\alias{sA}
\alias{sA,BAf-method}
\alias{sA<-}
\alias{sA<-,BAf-method}
\alias{codebook}
\alias{codebook,BAf-method}
\alias{codebook<-}
\alias{codebook<-,BAf-method}
\alias{note}
\alias{note,BAf-method}
\alias{note<-}
\alias{note<-,BAf-method}
\title{Access data stored in a BAf object}
\usage{
sI(x, ...)

\S4method{sI}{BAf}(x)

sI(x, ...) <- value

\S4method{sI}{BAf}(x) <- value

sX(x, ...)

\S4method{sX}{BAf}(x, b_id, drop = TRUE)

sX(x, ...) <- value

\S4method{sX}{BAf}(x) <- value

sB(x, ...)

\S4method{sB}{BAf}(x)

sB(x, ...) <- value

\S4method{sB}{BAf}(x) <- value

sA(x, ...)

\S4method{sA}{BAf}(x, wise = c("sinfo", "binder"))

sA(x, ...) <- value

\S4method{sA}{BAf}(x, wise = c("sinfo", "binder")) <- value

codebook(x, ...)

\S4method{codebook}{BAf}(x, wise = c("sinfo", "binder"))

codebook(x, ...) <- value

\S4method{codebook}{BAf}(x, wise = c("sinfo", "binder")) <- value

note(x)

\S4method{note}{BAf}(x)

note(x) <- value

\S4method{note}{BAf}(x) <- value
}
\arguments{
\item{x}{an object of the \code{\link{BAf-class}}}

\item{...}{for other functions with same name}

\item{value}{a \code{\link{tbl_df}}, \code{\link{data.frame}} or
\code{\link{matrix}} that fits to the accessing slot}

\item{b_id}{one or more binder IDs. This allows to find the data of those
binders ignoring repeat indicator, like \code{sX(x)[, match(b_id, bid(x,
exact = F))]}. For an exact match, use \code{`[`} as \code{sX(x)[, b_id]},
instead.}

\item{drop}{same as \code{\link{[}}}

\item{wise}{which batch, sample-wise or binder-wise.}
}
\value{
\tabular{ll}{
   \code{sI}\tab a \code{\link{tbl_df}} of sample information of the given
   BAf\cr 
   \code{sX}\tab a \code{\link{matrix}} or \code{\link{vector}} of numerical
   measurement values obtained from binder arrays\cr 
   \code{sB}\tab a \code{\link{data.frame}} of the information about
   binders\cr 
   \code{sA}\tab a \code{list} of \code{\link{data.frame}} of batch (or 
   assay) dependent information for the samples or binders\cr
   \code{codebook}\tab a \code{list} of two \code{\link{data.frame}}s named 
   as \code{sinfo} and \code{binder} unless \code{wise} is given. Each
   contains the codebook for the \code{@sinfo} or \code{@binder} above.\cr
   \code{note}\tab any note in any form
}
}
\description{
Short commands to access the information stored in a \code{\link{BAf-class}}.
\code{sI}, \code{sX}, \code{sB}, \code{sA}, \code{codebook}, and \code{note}
invoke the stored information in \code{@sinfo}, \code{@.Data},
\code{@binder}, \code{@assy_s} or \code{@assy_b}, \code{@codebook}, and
\code{note}, respectively. The details of the contents can be found in
\link{BAf-class}. The corresponding assignment functions e.g. \code{sI<-},
\code{sX<-}, \code{sB<-}, are also available.
}
\examples{
data(sba)

sI(sba)

sX(sba)[1:5, 1:5]           # (entire matrix)[1:5, 1:5]
sX(sba, "T2")[1:10]         # only the values obtained by 'T2' binder [1:10]
dim(sX(sba, "T2", drop= FALSE))

sB(sba)

head(sA(sba, "s")$fail_flag)
sA(sba, "sinfo")$fail_flag[1:5, ]
sA(sba, "sample")$fail_flag[1, ] <- FALSE      # 'sample' instead of 'sinfo'
sA(sba, "sinfo")$fail_flag[1:5, ]
head(sA(sba, "binder")$fail_flag)

head(codebook(sba))              # both codebooks for samples & binders
codebook(sba, "bin")         # short name
codebook(sba, "sample")      # 'sample' instead of 'sinfo'

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
