% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_colors_s_b.R
\name{default_color_SBA_bead}
\alias{default_color_SBA_bead}
\title{Default color of SBA beads}
\usage{
default_color_SBA_bead(only_ctrl)
}
\arguments{
\item{only_ctrl}{whether the colors of control antibodies only will be
returned or not}
}
\value{
a \code{\link{tibble}} that has \code{'name'}, \code{'col'},
  \code{'reg'}, and \code{'type'}
}
\description{
Returns the tibble for coloring binders
}
