% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_colors_s_b.R
\name{default_color_SBA_ctrl_samples}
\alias{default_color_SBA_ctrl_samples}
\title{Default colors of control samples}
\usage{
default_color_SBA_ctrl_samples(
  sampleid,
  rpt_id = sampleid[grep("^MIX_1", sampleid)]
)
}
\arguments{
\item{sampleid}{a vector of sample IDs}

\item{rpt_id}{The ID of replicated samples, the data of which is analyzed to 
check the stability of signals. Often pools are used, whose IDs start with 
"MIX_1". If missing, the IDs with the prefix "MIX_1" are classified as
replicated samples}
}
\value{
a \code{\link{tibble}} that has 'name', 'col', and 'reg'
}
\description{
The default colors of positive and negative control samples
}
\examples{
data(sba)
default_color_SBA_ctrl_samples(sampleid= sid(sba))

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
