% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.divideBy.R
\name{divideBy}
\alias{divideBy}
\title{divide by a column in @sinfo}
\usage{
divideBy(baf, by, abbr_name = FALSE)
}
\arguments{
\item{baf}{an object of the \code{\link{BAf-class}}}

\item{by}{the name of the column by which sample will be divided}

\item{abbr_name}{if the name of the list element should be shortened}
}
\value{
A \code{list} of \code{\link{BAf-class}} objects
}
\description{
It produces a list of BAf objects divided by the values in the given column
of \code{@sinfo}
}
\examples{
data(sba)
# the data "sba" is divided based on the values in "plate" column.
sba2 <- divideBy(sba, "plate")
str(sba2, 2)		# show the structure of "B"

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
