% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC_0_underlying.R
\name{.reorder_BAf_and_replace_0}
\alias{.reorder_BAf_and_replace_0}
\title{reorder sample and replace 0}
\usage{
.reorder_BAf_and_replace_0(
  baf,
  sample_order = order(batch(baf, "sample")),
  show_count_of_0 = FALSE
)
}
\arguments{
\item{baf}{an object of the \code{\link{BAf-class}}}

\item{sample_order}{the order of sample in plots. The default is the order by
sample batch in order to easily check batch differences.}

\item{show_count_of_0}{whether the number of 0 should be displayed}
}
\value{
an object of the \code{\link{BAf-class}}
}
\description{
Reorder the samples of an BAf according to given \code{sample_order} and
replace 0 with 1 to avoid any error in log-scale plots
}
