% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.failed_to_NA.R
\name{failed_to_NA}
\alias{failed_to_NA}
\title{Failed sample/binder to NA}
\usage{
failed_to_NA(
  baf,
  wise = c("sinfo", "binder"),
  fail_flag = "fail_flag",
  failMark = "failed",
  show_count = TRUE,
  by
)
}
\arguments{
\item{baf}{an object of the \code{\link{BAf-class}}}

\item{wise}{sample-wise or binder-wise.}

\item{fail_flag}{the name of the data frame that contains fail flags in
\code{@assy_s} or \code{@assy_b}}

\item{failMark}{what is the label in the flag for failed sample}

\item{show_count}{Whether the number of "failed" is shown}

\item{by}{if show_count is TRUE, then "by" indicates by which column the 
counts are stratified, e.g. "plate", "assay". This can be given as a vector
like the result of \code{batch(baf, "sinfo")}, which is the default.}
}
\value{
an object of the \code{\link{BAf-class}}
}
\description{
replace the data points of the failed samples/binders with NA
}
\examples{
data(sba)
failed_to_NA(sba, wise= "sinfo", by= "plate")
failed_to_NA(sba)         # default 'by' is batch

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
