% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.hist_correlation.R
\name{hist_correlation}
\alias{hist_correlation}
\title{Histogram of correlation}
\usage{
hist_correlation(
  baf,
  rep_binder,
  method = c("spearman", "pearson", "kendall"),
  use = "complete.obs",
  ...
)
}
\arguments{
\item{baf}{an object of the \code{\link{BAf-class}} or a matrix}

\item{rep_binder}{a \code{factor} (or a \code{vector}) that indicates which
binders were repeated, so their values should be combined.}

\item{method, use}{same as those of \code{\link{cor}}, but different defaults}

\item{...}{parameters that passed over to \code{\link{hist}}}
}
\value{
a \code{list} that has those correlation values
}
\description{
Histogram of correlation values between repeated measures, together with a
density plot of those between random pairs of measures
}
\seealso{
\code{\link{cor}}
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
