% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.BAf-class_key_ids.R
\docType{methods}
\name{key_id_access}
\alias{key_id_access}
\alias{sid}
\alias{sid,BAf-method}
\alias{sid<-}
\alias{sid<-,BAf-method}
\alias{bid}
\alias{bid,BAf-method}
\alias{bid<-}
\alias{bid<-,BAf-method}
\title{Access key ids}
\usage{
sid(x, exact = FALSE)

\S4method{sid}{BAf}(x, exact = FALSE)

sid(x) <- value

\S4method{sid}{BAf}(x) <- value

bid(x, exact = FALSE)

\S4method{bid}{BAf}(x, exact = FALSE)

bid(x) <- value

\S4method{bid}{BAf}(x) <- value
}
\arguments{
\item{x}{an object of the \code{\link{BAf-class}}}

\item{exact}{whether the exact ID including repeat indicator, i.e. \code{*1},
should be returned or not}

\item{value}{a \code{\link{vector}} of key ids}
}
\value{
\tabular{ll}{
   \code{sid}\tab a \code{vector} of sample IDs\cr
   \code{bid}\tab a \code{vector} of binder IDs\cr
}
}
\description{
\code{sid} returns a vector of sample IDs.\cr
\code{sid<-} replaces it with new \code{value}.

\code{bid} returns a vector of binder IDs.\cr
\code{bid<-} replace it with new \code{value}.
}
\examples{
data(sba)

head(sid(sba))

sid(sba)[1] <- "sA"

bid(sba)[1:10]

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
