% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.meanBy.R
\name{meanBy}
\alias{meanBy}
\title{Mean within subset}
\usage{
meanBy(baf, by, geometric = FALSE)
}
\arguments{
\item{baf}{an object of the \code{\link{BAf-class}}}

\item{by}{the column in \code{@sinfo} by which samples will be stratified}

\item{geometric}{if the geometic mean should be calculated}
}
\value{
A \code{"matrix"} of which the columns are levels in the columns of 'by' and
the rows are targets
}
\description{
Calculate an average of a target within the subset of samples divided by the
specified column in \code{@sinfo}
}
\examples{
data(sba)
meanBy(sba, "plate")		# average by "plate"
meanBy(sba, "dis", geometric = TRUE)		# geometric mean by "dis"

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
