% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_FlexMAP_csv.R
\name{parse_FlexMAP3D_csv_2_list}
\alias{parse_FlexMAP3D_csv_2_list}
\title{parse raw file from FLEXMAP3D}
\usage{
parse_FlexMAP3D_csv_2_list(fn)
}
\arguments{
\item{fn}{name of a file from FLEXMAP3D}
}
\value{
a \code{list}, the elements of which are
  \item{head}{\code{'data.frame'} that contains information about the
  experiment, such as date, instrument id, etc}
  \item{pos}{\code{'data.frame'} indicates the position on plate}
  \item{other}{The corresponding data in \code{'data.frame'}}
}
\description{
The function to parse raw file attained directly from the FLEXMAP3D
instrument
}
\seealso{
\code{\link{read_LIMS_SBA_files}}
\code{\link{read_FlexMAP3D_csv}}
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
