% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateLayout_0.6.R
\name{plateLayout.SBA}
\alias{plateLayout.SBA}
\title{Design plate layout}
\usage{
plateLayout.SBA(sinfo, cat_var = NULL, qty_var = NULL,
  nNull_per_96 = 2, null_var = NULL, plate_nr = NULL)
}
\arguments{
\item{sinfo}{a data frame that includes information about samples}

\item{cat_var}{categorical variables into consideration. The number of 
samples in each category will be balanced (or kept constant) across the 
plates. It should be a value or a vector of some of column names in 
\var{sinfo}. When the number of samples are not a multiple of number of
plates, some variation in numbers cannot be avoided. In such case, the
variables listed to the left are prioritized during the balancing.}

\item{qty_var}{quantitative variables into consideration. The values of the 
variables will be tested by ANOVA to check the balance and displayed in box
plots. Like \var{cat_var}, it should be in \var{sinfo}}

\item{nNull_per_96}{the number of null wells on each 96-well plate}

\item{null_var}{the name of the column in which 'NULL' will be shown to 
indicate null well in \var{sinfo}. If it is given to NULL as default, the
rows for null well will not be appended.}

\item{plate_nr}{the plate number(s) to which samples will be allocated. The 
default is a sequential number.}
}
\value{
The extended data frame of \var{sinfo} is generated with the columns
\code{'plate'}, \code{'pos'}, \code{'row'}, and \code{'col'}. \code{'plate'}
is designated plate ID. \code{'pos'} is the number indicating position within
plate (from top left to bottom right, move column-wise first). \code{'row'}
and \code{'col'} are the row letter and column number, respectively. The
plots to assist to compare the distribution of quantitative variables in each
plate are accompanied.
}
\description{
Design layouts of upto four 96-well plates by randomization with balances 
across those plates in terms of potential confounders. This can handle more
than 384 samples, then it produces layouts for plural 386-well plates.
}
\details{
- Balancing in the randomization are as below,\cr
  1) The number of samples of each group devided by 'categorical variables'
  are same or similar on each plate
    e.g. Male patients on each plate are same across all plates. It lets the
    ratios of male:female and patient:control are kept.
  2) The means of the values of 'quantitative variables' are homogeneous
  across all plates. 
    e.g. The mean of ages of a plate is not too much deviated from the mean
    of other plates.
}
\examples{

n <- 368
tg <- data.frame(
  "id" = c(paste("a", sample(n*2, n), sep="_"), rep("repl", 8)), 
  "age" = c(abs(rnorm(n, 50, 20)), rep(NA, 8)), 
  "diag_yrs" = c(abs(rnorm(n, 2)), rep(NA, 8)), 
  "gender" = c(rep(c("female", "male"), n/2), rep(NA, 8)), 
  "dis" = c(rep(c("case", "control"), each = n/2), rep("repl", 8))
)
X <- plateLayout.SBA(tg, c("dis", "gender"), c("age", "diag_yrs"), 
                    null_var= "id", plate_nr= 1:4)
head(tg)
head(X)

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
