% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_a_pair_vectors.R
\name{plot_a_pair_vectors}
\alias{plot_a_pair_vectors}
\title{Plot to compare a pair of vectors}
\usage{
plot_a_pair_vectors(x, y, grp_id, square = TRUE, log = "xy",
  main = paste("Scatter plot of", xlab, "&", ylab),
  xlab = deparse(substitute(x)), ylab = deparse(substitute(y)),
  col_ea_grp, ..., legend_arg = list(x = "bottomright"))
}
\arguments{
\item{x}{a vector of numerical values}

\item{y}{a vector of values that are very likely correlated with values in 'x'}

\item{grp_id}{a vector that has the group IDs of which each will be plotted
in different color\cr - It should have same length as \code{x} and \code{y}}

\item{square}{whether the ranges of the x- and y-axes in plot should same or
not}

\item{log}{which coordinates are in log scale. Same as \code{'log'} in 
\code{\link{plot.default}}.}

\item{main, xlab, ylab}{same as the ones for \code{'plot'} generic function 
(refer \code{\link{par}})}

\item{col_ea_grp}{color of each group given in \code{'grp_id'}}

\item{...}{any arguments that pass to \code{'plot'} and \code{'points'}}

\item{legend_arg}{a list of parameters to be used in adding a legend. If 
\code{grp_id} is missing then no legend will be displayed.}
}
\value{
A plot is generated and a \code{"list"} of x- and y- coordinate
  limits is returned.
}
\description{
The plot function to compare a pair of vectors. One of the main usages is to
check how homogeneous is the data generated on multiple plates. The points
are colored by group.
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
