% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_a_pair_vectors.R
\name{plot_a_pair_vectors_in_MA}
\alias{plot_a_pair_vectors_in_MA}
\title{MA-plot with coloring by group}
\usage{
plot_a_pair_vectors_in_MA(
  x1,
  x2,
  grp_id,
  x1name = deparse(substitute(x1)),
  x2name = deparse(substitute(x2)),
  log = TRUE,
  main,
  xlab,
  ylab,
  ...,
  legend_arg
)
}
\arguments{
\item{x1, x2}{the first and second vectors to show}

\item{grp_id}{a vector that has the group IDs of which each will be plotted
in different color\cr - It should have same length as \var{x1} and \var{x2}}

\item{x1name, x2name}{the names of \var{x1} and \var{x2}. These will be used
in the labels of x-, y-axis}

\item{log}{whether the \code{MA} should be computed in log-scale or not. If
this is TRUE, 'x' coordinate is displayed in log scale}

\item{main}{same as the ones for \code{'plot'} generic function 
(refer \code{\link{par}})}

\item{xlab}{same as the ones for \code{'plot'} generic function 
(refer \code{\link{par}})}

\item{ylab}{same as the ones for \code{'plot'} generic function 
(refer \code{\link{par}})}

\item{...}{any arguments that pass to \code{plot_a_pair_vectors}}

\item{legend_arg}{a list of parameters to be used in adding a legend. If 
\code{grp_id} is missing then no legend will be displayed.}
}
\value{
A plot is generated and a \code{"list"} of x- and y- coordinate
  limits is returned.
}
\description{
MA-plot with coloring by group
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
