% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.BAf.R
\docType{methods}
\name{Plot_of_mean_comparison}
\alias{Plot_of_mean_comparison}
\alias{plot_compareMeanOfEach_inPairs}
\alias{plot_compareMeanOfEach_inPairs,matrix-method}
\alias{plot_compareMeanOfEach_inPairs,BAf-method}
\title{To compare means in every plates}
\usage{
plot_compareMeanOfEach_inPairs(X, ...)

\S4method{plot_compareMeanOfEach_inPairs}{matrix}(X, groupID, log = TRUE,
  MAplot = FALSE, main, legend_par = list(x = "bottomright"), ...)

\S4method{plot_compareMeanOfEach_inPairs}{BAf}(X, geometric = FALSE,
  by = "plate", log = TRUE, MAplot = FALSE,
  main = paste0("Comparing", if (geometric) " geometric" else "",
  " means"), legend_par = list(x = "bottomright"), prefix_plate = by,
  ...)
}
\arguments{
\item{X}{a \code{\link{BAf}} object or a matrix in which each ROW represents each TARGET and each
COLUMN indicates each COLLECTION of samples to be paired. The column name should indicate the
name of the collections.}

\item{...}{any arguments that pass to \code{\link{plot}} and \code{\link{points}}}

\item{groupID}{a vector of the group ids. Each group will be plotted in different color. It
should have same number of rows of \code{X}}

\item{log}{logically indicates if the coordinates are in log scale}

\item{MAplot}{whether an MA-plot is wanted}

\item{main}{same as the ones for \code{'plot'} generic function (refer \code{\link{par}})}

\item{legend_par}{a list of parameter to be used in adding a legend. If \code{groupID} is missing
then no legend will be added.}

\item{geometric}{if the geometric mean is computed instead of arithmetic one.}

\item{by}{by which column the samples should be devided during calculation of the mean}

\item{prefix_plate}{the prefix added to the plate ID shown in X- and Y-axes}
}
\value{
A plot is generated and a \code{"list"} of x- and y- coordinate limits is returned.
}
\description{
plot to observe correlation between the estimates (e.g. mean) that represent intensities for each
antibody in a single window coloring each assay separately.
}
\examples{
data(sba)
\dontrun{plot_compareMeanOfEach_inPairs.BAf(sba)}

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
