% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PC.R
\name{plot_PC}
\alias{plot_PC}
\alias{plot_PC,prcomp-method}
\alias{plot_PC,matrix-method}
\alias{plot_PC,BAf-method}
\title{Plot of principal components}
\usage{
plot_PC(x, pc = c(1, 2), main = "", xlab, ylab, ...)

\S4method{plot_PC}{prcomp}(x, pc = c(1, 2), main = "", xlab, ylab, ...)

\S4method{plot_PC}{matrix}(x, pc = c(1, 2), main = "", xlab, ylab, ...)

\S4method{plot_PC}{BAf}(
  x,
  pc = c(1, 2),
  main = "",
  xlab,
  ylab,
  ...,
  col_by = NULL,
  pch_by = NULL,
  legend_arg = list(x = "topright")
)
}
\arguments{
\item{x}{an object of the \code{\link{prcomp}}, \code{\link{matrix}} or
\code{\link{BAf-class}}}

\item{pc}{a vector that consists of two integer elements for two principal
components}

\item{main, xlab, ylab}{same as \code{\link{plot.default}}}

\item{...}{pass to \code{plot}}

\item{col_by, pch_by}{a name of a column in \code{sI(x)}. Different color /
symbols the points will have.}

\item{legend_arg}{a list of parameters to be used in adding a legend. If
NULL, skip to add legend}
}
\value{
\code{plot_PC} \code{\link{prcomp}} output object
}
\description{
The plot of a pair of principal components is displayed for a matrix or a
single \code{\link{BAf-class}} object. Colors or symbols of the points can be
diffrentiated by the variable given in \var{col_by} or \var{pch_by} to assist
to find hidden substructure of sample. The principal components were obtained
by (\code{\link{prcomp}}) with scaling.
}
\section{Functions}{
\itemize{
\item \code{plot_PC,prcomp-method}: The input is the output of 'prcomp'

\item \code{plot_PC,matrix-method}: The input is a 'matrix'

\item \code{plot_PC,BAf-method}: The input is a 'BAf' object
}}

\examples{
data(sba)
plot_PC(sba)
plot_PC(sba, col_by= "plate")    # no legend
plot_PC(sba, col_by= "plate", pch_by= "sex", legend_arg= list())

}
\seealso{
\code{\link{prcomp}}
}
\author{
Mun-Gwan Hong \email{mun-gwan.hong@scilifelab.se}
}
