% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC.SBA.R
\name{plot_QC_bead_count_boxplot}
\alias{plot_QC_bead_count_boxplot}
\title{Bead count plot}
\usage{
plot_QC_bead_count_boxplot(bead_count, sample_order = NULL,
  color_tbl_b = default_color_SBA_bead(T),
  color_tbl_s = default_color_SBA_ctrl_samples(sampleid),
  main = deparse(substitute(bead_count)), xlab = c("Beads",
  "Sample (sorted)"), ylab = "Bead count", ylim, ...)
}
\arguments{
\item{bead_count}{1) a \code{data.frame} of bead counts having sample IDs at
the first column like the file from LIMS, 2) a file name that contains such
data, or 3) a \code{matrix} of such values only where sample IDs are in
\code{rownames}}

\item{sample_order}{the order of sample in plots. The default is same order 
as given table by \var{bead_count}.}

\item{color_tbl_b}{the colors of the binders employed for positive or 
negative control on plot. The information should be provided as a 
\code{\link{data.frame}}, in which columns are \code{name}, \code{reg}
(regular expression to find the binder), \code{col} (color), and
\code{type} (\code{positive} , \code{negative}, or \code{measure}). If this
is given as NULL, it results in no coloring.}

\item{color_tbl_s}{colors of samples on the plot. The information should be 
provided as a \code{\link{data.frame}} that has \code{name}, \code{reg} 
(regular expression to find the sample), \code{col} (color), \code{type} 
(\code{negative} or \code{replicated}).\cr If this is NULL, then no 
coloring is applied. The default is
\code{\link{default_color_SBA_ctrl_samples}(sampleid)}, where
\code{sampleid} is a vector of sample IDs that are extracted from
\var{bead_count}}

\item{main, ylab, ylim, ...}{same as those for \code{\link{boxplot}}}

\item{xlab}{two characters. The first one is for the plot of binders, while
the 2nd is for sample plot}
}
\value{
a \code{list} of \code{color_tbl_b} and \code{color_tbl_s}
}
\description{
Display bead counts across sample and binders.
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
