% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC_binder.R
\name{plot_QC_binder_signal_boxplot}
\alias{plot_QC_binder_signal_boxplot}
\alias{plot_QC_binder_signal_boxplot.SBA}
\title{Signal distribution across binders}
\usage{
plot_QC_binder_signal_boxplot(baf, color_tbl_b = NULL,
  main = deparse(substitute(baf)), xlab = "Binders",
  ylab = "Signal Intensity", guide_line = TRUE, ...)

plot_QC_binder_signal_boxplot.SBA(baf,
  color_tbl_b = default_color_SBA_bead(T),
  main = paste(unique(batch(baf, "binder")), collapse = ", "),
  xlab = "Beads", ylab = "Raw MFI", guide_line = TRUE, ...)
}
\arguments{
\item{baf}{an object of the \code{\link{BAf-class}}}

\item{color_tbl_b}{the colors of the binders employed for positive or 
negative control on plot. The information should be provided as a 
\code{\link{data.frame}}, in which columns are \code{name}, \code{reg}
(regular expression to find the binder), \code{col} (color), and
\code{type} (\code{positive} , \code{negative}, or \code{measure}). If this
is given as NULL, it results in no coloring.}

\item{main, xlab, ylab, ...}{same as \code{\link{boxplot}}}

\item{guide_line}{if some guide lines of controls are displayed.}
}
\value{
a \code{list} of \code{color_tbl_b}
}
\description{
Depict a plot that displays signal distribution across samples
}
\section{Functions}{
\itemize{
\item \code{plot_QC_binder_signal_boxplot.SBA}: when \var{baf} has the data from
suspension bead arrays
}}

\examples{
data(sba)
plot_QC_binder_signal_boxplot(sba)

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
