% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC_repl_var.R
\name{plot_QC_repl_var}
\alias{plot_QC_repl_var}
\title{Replicates variation plot}
\usage{
plot_QC_repl_var(
  baf,
  i_repl = 1:nrow(baf),
  by_s = batch_colname(baf, "sample"),
  varFUN = function(x) {     sd(x, na.rm = T)/mean(x, na.rm = T) },
  col_lines = palette(),
  main = deparse(substitute(baf)),
  xlab = "Coef. of var.",
  ylab = "Density",
  xlim = c(0, 0.5),
  ylim = NULL,
  lty = c("solid", "dashed"),
  lwd = c(2, 1),
  ...
)
}
\arguments{
\item{baf}{an object of the \code{\link{BAf-class}}}

\item{i_repl}{an integer or logical index vector of the rows that have
replicated samples}

\item{by_s}{by which the samples are grouped. The \var{varFUN} is applied to 
each group and the results are combined. This should be given by a 
\code{\link{character}} string of a column name in \code{@sinfo}.}

\item{varFUN}{a function to compute a variation statistics. The default is
the function to compute the coefficienct of variation.}

\item{col_lines}{a named vector of line colors in character strings}

\item{main, xlab, ylab, xlim, ylim}{same as those for \code{\link{plot}}}

\item{lty, lwd}{the first value is for the line to show the variation of
replicates, while the second is for the other samples. Individual values
are same those for \code{\link{plot}}}

\item{...}{the arguments passed over to \code{plot} and \code{lines}}
}
\value{
a list that contains computed varaition statistics of the replicates
  and all the other samples
}
\description{
Draw the plot that shows the distribution of variation statistics (e.g.
coefficient of variation(CV)) of replicates
}
\examples{
data(sba)
sba2 <- sba[sba@sinfo$cohort != "EMPTY", ]
plot_QC_repl_var(sba2, sba2@sinfo$cohort == "MIX_1")

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
