% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC_sample.R
\name{plot_QC_sample_signal_boxplot}
\alias{plot_QC_sample_signal_boxplot}
\alias{plot_QC_sample_signal_boxplot.SBA}
\title{Signal distribution across samples}
\usage{
plot_QC_sample_signal_boxplot(baf, sample_order = order(batch(baf,
  "sample")), color_tbl_s = NULL, main = deparse(substitute(baf)),
  xlab = "Sample (sorted)", ylab = "Signal Intensity", ...)

plot_QC_sample_signal_boxplot.SBA(baf, sample_order = order(batch(baf,
  "sample")), color_tbl_s = default_color_SBA_ctrl_samples(sid(baf)),
  main = paste(unique(batch(baf, "binder")), collapse = ", "),
  xlab = "Sample (sorted)", ylab = "Raw MFI", ...)
}
\arguments{
\item{baf}{an object of the \code{\link{BAf-class}}}

\item{sample_order}{the order of sample in plots. The default is the order by
sample batch in order to easily check batch differences.}

\item{color_tbl_s}{colors of samples on the plot. The information should be 
provided as a \code{\link{data.frame}} that has \code{name}, \code{reg} 
(regular expression to find the sample), \code{col} (color), \code{type} 
(\code{negative} or \code{replicated}).\cr If this is NULL, then no
coloring is applied.}

\item{main, xlab, ylab, ...}{same as \code{\link{boxplot}}}
}
\value{
a \code{list} of the "\var{color_tbl_s}"
}
\description{
Draw a plot that displays signal distribution across samples
}
\section{Functions}{
\itemize{
\item \code{plot_QC_sample_signal_boxplot.SBA}: when \var{baf} has the data from
suspension bead arrays
}}

\examples{
data(sba)
plot_QC_sample_signal_boxplot(sba)

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
