% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_PQN.R
\docType{methods}
\name{pqn}
\alias{pqn}
\alias{pqn,BAf-method}
\title{Probabilistic Quotient Normaliation}
\usage{
pqn(X, preNorm = FALSE, ...)

\S4method{pqn}{BAf}(X, preNorm = FALSE, by_s = NULL, by_b = NULL,
  ...)
}
\arguments{
\item{X}{a \code{\link{matrix}} or a \code{\link{BAf-class}} object}

\item{preNorm}{if the integral normalization should be applied prior to the
PQN. Please refer to the Dieterle \emph{et al.}'s paper.}

\item{...}{not used}

\item{by_s, by_b}{same as \code{by_s} and \code{by_s} in
\code{\link{apply_per_group}}. If any of these are given the \code{pqn} is
applied per group divided by these variables.}
}
\value{
\code{\link{matrix}} (or \code{\link{BAf-class}} object) after the
  normalization
}
\description{
Normalize the data applying Probabilistic Quotient method introduced in
Dieterle et al.(2006).
}
\examples{
data(sba)
sba2 <- pqn(sba[sba@sinfo$cohort != "EMPTY", ])
plot_QC_sample_signal_boxplot(sba2)

}
\references{
Dieterle et al. (2006) Probabilistic quotient normalization.. - Anal.Chem
}
\seealso{
\code{\link{apply_per_group}}
\code{\link{normn_MA}}
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
