% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_FlexMAP_csv.R
\name{read_FlexMAP3D_csv}
\alias{read_FlexMAP3D_csv}
\title{Generate BAf object parsing raw file from FLEXMAP3D}
\usage{
read_FlexMAP3D_csv(file, assayid, sinfo, bead, sample_batch_c = "plate",
  sbaid, bead_count_as_failed = 20, bead_count_alert = 35,
  reading_order = 1:nrow(sinfo), QC_plot = FALSE,
  path_QC = "../output/QC", i_neg_ctrl_sample, row_var, col_var)
}
\arguments{
\item{file}{name of a file from FLEXMAP3D}

\item{assayid}{the assay ID of this sample set and antibody set.}

\item{sinfo}{a \code{tbl_df} or \code{data.frame} that contains information
about samples, in which each row is for each sample}

\item{bead}{a \code{tbl_df} or \code{data.frame} having bead information. 
Often it can be obtained from other BAf object. If this is missing, a
simple \code{tbl_df} that has only 'assay' column will be used as default.}

\item{sample_batch_c}{the name of sample batch column in the \code{sinfo}.
e.g. \code{plate}}

\item{sbaid}{optionally, SBA ID of this antibody set can be given here.}

\item{bead_count_as_failed, bead_count_alert}{the cutoff values for bead
count. Labeled in \code{'fail_flag'} as \code{'failed'} if the bead count
is lower than \code{bead_count_as_failed}. If it is below
\code{bead_count_alert}, then marked as \code{'lowBeadCount'}. Not in both
cases, it is marked as \code{'ok'}.}

\item{reading_order}{the integers in which order the sample has read by
FlexMAP3D. This parameter allows to accept the experimental data which is
conducted in reverse or random order.}

\item{QC_plot}{if TRUE, some plots for quality check-up are generated like
that of \code{\link{read_LIMS_SBA_files}}}

\item{path_QC}{where the QC plots will be saved. This is valid only if
\code{QC_plot} is TRUE.}

\item{i_neg_ctrl_sample}{the index of negative control samples such as
\code{EMPTY}. A logical vector of same length as samples can be given.}

\item{row_var, col_var}{When these are given, the variable in \code{sinfo} is
compared with the values of row or column in \code{file} to confirm the
order of samples in \code{sinfo}.}
}
\value{
an object of \link{BAf-class}
}
\description{
This generates an BAf object parsing raw file attained directly from the
FLEXMAP3D instrument. The samples with median bead counts lower than 35 are
marked as "failed" in \code{@assay$fail_flag}
}
\seealso{
\code{\link{read_LIMS_SBA_files}}
\code{\link{parse_FlexMAP3D_csv_2_list}}
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
