% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_LIMS_SBA_files.R
\name{read_LIMS_SBA_files}
\alias{read_LIMS_SBA_files}
\title{Parse a set of constituent files exported from LIMS for a single project.}
\usage{
read_LIMS_SBA_files(SLid, path, ..., QC_plot = FALSE,
  QC_report = FALSE, author = "Mun-Gwan", path_QC = "../QC",
  lowbound_mfi = 100)
}
\arguments{
\item{SLid}{the sample layout ID used in LIMS, e.g. "SL0001". Currently it is
the first six letters in all constituent files.}

\item{path}{the path to the files from LIMS}

\item{...}{for the future}

\item{QC_plot}{whether the plots in technical aspects should be drawn, such
as intentisy distribution plots across samples or antibodies.}

\item{QC_report}{whether a primary QC report to be written invoking
\code{\link{write_primary_QC_report.SBA}} or not}

\item{author}{the author of the report}

\item{path_QC}{where the QC plots or QC report to be saved}

\item{lowbound_mfi}{the lower bound of MFI. If a median of a sample is below
this threshold, the data of the sample is neglected afterwords in QC
report.}
}
\value{
an object of the \code{\link{BAf-class}}
}
\description{
A set of files for a project can be obtained from "Project Summary" in LIMS.
This function reads the set and construct an object of BAf class parsing the
files. Because this function is so sensitive to the format of data, the
functionality of the current version of this function can limited to
present-day form (2014-03-04) only with limited backward compatibility.
Please, note that any change in LIMS can affect this function. Optionally, it
produces some plots in technical aspects.
}
\seealso{
\code{\link{read_FlexMAP3D_csv}}
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
