% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Olink_excel.R
\name{read_Olink_excel}
\alias{read_Olink_excel}
\title{Read Olink Excel}
\usage{
read_Olink_excel(file)
}
\arguments{
\item{file}{name of a file from Olink}
}
\value{
an object of \link{BAf-class}
}
\description{
This generates an BAf object parsing an output file in Excel format from
Olink instrument
}
\seealso{
\code{\link{read_excel}}
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
