% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.BAf-class_key_ids.R
\name{repeat_indicator}
\alias{repeat_indicator}
\alias{rm_repeat_indicator}
\alias{make.unique_key_ids}
\title{Handle repeat indicator of the key IDs}
\usage{
rm_repeat_indicator(ids)

make.unique_key_ids(ids)
}
\arguments{
\item{ids}{a \code{vector} of the key IDs (\code{'id'} in \code{@sinfo} or
\code{@binder}). It is by default converted to \code{character}.}
}
\value{
\code{rm_repeat_indicator} A \code{vector} of \code{character} of
  the 'id' after removal of the indicator

\code{make.unique_key_ids} A \code{vector} of \code{character} of 
  cleaned 'id's with the suffix
}
\description{
The key IDs (\code{'id'} in \code{@sinfo} or \code{@binder}) should be unique
throughout the samples and binders, in order to distinguish among them. So,
any \code{id}s repeatedly used are amended by adding "*" and the digit
indicating how many times the same ids have been employed ahead as a suffix
(e.g. HPA0001, HPA0001*1)
\code{rm_repeat_indicator} - Remove the additional suffix (repeat indicator)

\code{make.unique_key_ids} generate unique 'id's by adding the suffices. When
necessary, the indicators previously attached are removed first.
}
\examples{
rm_repeat_indicator(c("HPA001", "HPA001*1"))

make.unique_key_ids(c("HPA001", "HPA003", "HPA001", "HPA001", "HPA003"))

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
