% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.replace_x.R
\name{replace_0}
\alias{replace_0}
\title{replace 0 with given value}
\usage{
replace_0(baf, value = min(sX(baf)[sX(baf) > 1], na.rm = TRUE) - 1,
  show_count = TRUE, by_s = NULL, by_b = NULL)
}
\arguments{
\item{baf}{an object of the \code{\link{BAf-class}}}

\item{value}{the values with which zeros will be replaced}

\item{show_count}{whether the number appearance of 0s is displayed or not}

\item{by_s, by_b}{if show_count is TRUE, then these specify by which column the 
counts are stratified, e.g. "plate", "assay".}
}
\description{
replace 0 with a given value, The default is (the lowest value - 1)
}
\examples{
data(sba)
replace_0(sba)

}
