% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sba.R
\docType{data}
\name{sba}
\alias{sba}
\title{data - sba}
\format{an object of the class \code{\link{BAf-class}}}
\usage{
data(sba)
}
\description{
The data given in this pacakge for examples. It contains artificially 
generated data that mimick the data from Suspension bead arrays.
}
\examples{

data(sba)
summary(sba)
}
\keyword{datasets}
