% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sia.standard_curve_function.R
\name{std_curve}
\alias{std_curve}
\title{Standard curve function}
\usage{
std_curve(std, std_conc, ...)
}
\arguments{
\item{std}{the obtained signals of standard dilution series}

\item{std_conc}{the known concentration of the standard series}

\item{...}{not used}
}
\value{
a list of three elements. The first, \code{fn} has a function that
  compute concentration when signal measure is given. Other two are
  \code{LOD} (limit of detection) and \code{LLOQ} (lower limit of
  quantification). The last is an \code{\link{drc}} object that has fitting
  model.
}
\description{
Using the data of standard dilution series, find the function that converts
obtained signal into concentration
}
\examples{
std <- c(17933, 8088, 2160, 373, 71, 9, 8, 17076, 8064, 2283, 401, 85, 8, 10)
std_conc <- rep(c(34589, 8647, 2161, 540, 135, 0, 0), 2)
scurv <- std_curve(std, std_conc)
scurv$fn(386)

}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
