% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.extract.R
\name{[,BAf,ANY,ANY,ANY-method}
\alias{[,BAf,ANY,ANY,ANY-method}
\title{Extract a part of data}
\usage{
\S4method{[}{BAf,ANY,ANY,ANY}(x, i, j, ..., droplevel = TRUE, drop)
}
\arguments{
\item{x}{an object of the \code{\link{BAf-class}}}

\item{i}{row}

\item{j}{column}

\item{..., drop}{currently not used}

\item{droplevel}{whether unused factor levels after extraction should be
dropped out in \code{@sinfo} and \code{@binder}}
}
\value{
an object of the \code{\link{BAf-class}} that contains a part
}
\description{
This function works in almost same way as the generic form of it.\cr
Note: When \code{j} is missing, this takes rows instead of columns.
}
\examples{
data(sba)
# the samples in 1:4 rows and the values of 3rd target
sba2 <- sba[1:4, 3]
summary(sba2)

# the data of "T2" and "T3"
sba3 <- sba[, c("T2", "T3")]
summary(sba3)

# the data of the samples on plate 2
sba4 <- sba[sba@sinfo$plate == "2", ]
summary(sba4)

# the data of cancer patients only
sba5 <- sba[!is.na(sba@sinfo$dis) & sba@sinfo$dis == "cancer", ]
summary(sba5)

}
\seealso{
\code{\link{Extract}}
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
