% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.summary.R
\name{summary}
\alias{summary}
\alias{summary,BAf-method}
\title{summary of the BAf object}
\usage{
summary(object, ...)

\S4method{summary}{BAf}(object, ...)
}
\arguments{
\item{object}{an object of the \code{\link{BAf-class}}}

\item{...}{not used}
}
\value{
a list of summarized elements
}
\description{
This shows the summary of \code{@sinfo}, \code{@binder}, and \code{@.Data}
}
\examples{
data(sba)
summary(sba)

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
