% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_measure.R
\name{transform_to_landscape_for_repeats}
\alias{transform_to_landscape_for_repeats}
\title{Transform to landscape form}
\usage{
transform_to_landscape_for_repeats(baf)
}
\arguments{
\item{baf}{an object of \code{\link{BAf-class}}}
}
\value{
\code{\link{BAf-class}} object in landscape form
}
\description{
When a \code{\link{BAf-class}} object contains the data of multiple repeats
in one assay, this function transform the data into landscape form, in which
the repeats are extended into columns.
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
