% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_position.R
\name{well_position}
\alias{well_position}
\alias{well_rowcol}
\alias{well_pos}
\alias{well_row384}
\alias{well_col384}
\title{Conversion of well position index}
\usage{
well_rowcol(x, row = TRUE, col = TRUE, as.factor = FALSE)

well_pos(rc, n_well = c(96, 384))

well_row384(pos, plate, as.factor = TRUE)

well_col384(pos, plate, as.factor = TRUE)
}
\arguments{
\item{x}{a vector of position numbers}

\item{row}{if the letters for rows should be returned}

\item{col}{if the column number should be returned}

\item{as.factor}{if the returned vector should be converted to factor}

\item{rc}{a vector of position letters composed of row letter and column
number e.g. B02}

\item{n_well}{the number of wells per plate. It should be 96 or 384. The
default is 96.}

\item{pos}{a vector of position number on 96-well plate}

\item{plate}{a vector of plate number of 96-well plates}
}
\value{
\code{well_rowcol} produces a character vector that has well position letter

\code{well_pos} an integer vector of position number

\code{well_row384} a factor or a vector of characters

\code{well_col384} a factor or a vector of characters
}
\description{
The functions to convert the index of well position in 96 or 384-well plate
\code{well_rowcol} - Convert position number to letter format.
}
\examples{
  well_rowcol(sample(1:96, 10))

    well_pos("B01")

}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
