% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_primary_QC_report.SBA.R
\name{write_primary_QC_report.SBA}
\alias{write_primary_QC_report.SBA}
\title{Write primary QC report}
\usage{
write_primary_QC_report.SBA(baf, bead_count = NULL,
  sample_order = order(batch(baf, "sample")),
  file = "primary_QC_report.html",
  title = paste("Primary QC report of", deparse(substitute(baf))),
  author = "Mun-Gwan", date = Sys.Date(), rm_Rmd = T,
  color_tbl_b = default_color_SBA_bead(T),
  color_tbl_s = default_color_SBA_ctrl_samples(sid(baf)),
  lowbound_mfi = 100)
}
\arguments{
\item{baf}{an object of the \code{\link{BAf-class}} that contains SBA data}

\item{bead_count}{1) a \code{data.frame}, 2) a file name or 3) a
\code{matrix}. check \code{\link{plot_QC_bead_count_boxplot}}}

\item{sample_order}{the order of sample in plots. The default is the order by
sample batch in order to easily check batch differences.}

\item{file}{The name of the generating HTML file. The ".html" extension will
be added if it is missing.}

\item{title}{The title of the generating report, which is displayed at the
top}

\item{author}{The author of the report}

\item{date}{The date written as the creation date}

\item{rm_Rmd}{if the .Rmd (RMarkdown) file should be removed after .pdf was
generated.}

\item{color_tbl_b}{the colors of the binders employed for positive or 
negative control on plot. The information should be provided as a 
\code{\link{data.frame}}, in which columns are \code{name}, \code{reg}
(regular expression to find the binder), \code{col} (color), and
\code{type} (\code{positive} , \code{negative}, or \code{measure}). If this
is given as NULL, it results in no coloring.}

\item{color_tbl_s}{colors of samples on the plot. The information should be 
provided as a \code{\link{data.frame}} that has \code{name}, \code{reg} 
(regular expression to find the sample), \code{col} (color), \code{type} 
(\code{negative} or \code{replicated}).\cr If this is NULL, then no
coloring is applied.}

\item{lowbound_mfi}{the lower bound of MFI. If a median of a sample is below
this threshold, the data of the sample is neglected afterwords in this QC
report.}
}
\description{
Write a primary QC report for SBA data using Rmarkdown
}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
