% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_textfiles_to_share.R
\name{write_textfiles_to_share}
\alias{write_textfiles_to_share}
\alias{write_textfiles_to_share.SBA}
\title{Text files to share}
\usage{
write_textfiles_to_share(
  baf,
  prefix = "BAf",
  suffix = paste0("(", Sys.Date(), ")"),
  path = "_sent",
  fn_data = "-data",
  fn_sample = NULL,
  fn_binder = NULL,
  incl_sample_i = c("id"),
  incl_binder_i = c("id"),
  fix_sid = c(),
  quote = TRUE,
  sep = "\\t"
)

write_textfiles_to_share.SBA(
  baf,
  prefix = "SBA",
  suffix = paste0("(", Sys.Date(), ")"),
  path = "_sent",
  fn_data = "-data",
  fn_sample = NULL,
  fn_binder = NULL,
  incl_sample_i = c("id"),
  incl_binder_i = c("id", "gene_name", "gene_description", "ensg_id"),
  fix_sid = c(NEG_CTRL = "EMPTY-0001", REPEATED_SAMPLE = "MIX_1-0001"),
  quote = TRUE,
  sep = "\\t"
)
}
\arguments{
\item{baf}{an object of the \code{\link{BAf-class}}, of which the data will
be converted to text files}

\item{prefix}{prefix used in file name. An ID for sample set is recommended.}

\item{suffix}{the suffix in the file name}

\item{path}{where those files should be saved.}

\item{fn_data}{the characters added into the name of the BAf MFI data file to
indicate it contains data (e.g '-data')}

\item{fn_sample, fn_binder}{the characters added into the name of the file of
sample/binder information. If this is NULL as the default, the file will be
not generated.}

\item{incl_sample_i, incl_binder_i}{the columns in BAf@sinfo/@binder that will
be included in 'data'/'binder' file. The column for sample IDs at least
should be included in \code{incl_sample_i}.}

\item{fix_sid}{a named vector of characters. The sample IDs that match to any
of those characters will be replaced with the corresponding name in this
vector. An examples is \code{c(NEG_CTRL = "EMPTY-0001", REPEATED_SAMPLE =
"MIX_1-0001")}}

\item{quote}{same as \code{quote} in \code{\link{write.table}}. Default is 
TRUE}

\item{sep}{seperator in the output text files. Tab is the default.}
}
\description{
Write text files to share the data stored in a \code{\link{BAf-class}} object
with collaborators
}
\section{Functions}{
\itemize{
\item \code{write_textfiles_to_share.SBA}: When 'baf' has SBA data. Just different
default settings.
}}

\seealso{
\code{\link{ask_write.table}}
}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
